Install the package so that we can execute the application on the command-line.
