+++ requirements.test.txt
@@ -1,0 +1,1 @@
+tox==2.8.1
