+++ tox.ini
@@ -6,5 +6,5 @@ skipsdist = True
 deps = -r{toxinidir}/requirements.txt
        coverage
        coveralls[yaml]
-commands = coverage run --source=rest_tester -m unittest discover
+commands = coverage run --source=rest_tester --omit=command_line.py -m unittest discover
            coveralls
